Update Makefile with conditions
The Makefile compiles according to whether the machine is
a sandybridge or not and whether icc is present or not.
If no icc, the Makefile compiles with gcc and clang_llvm.
If icc, the Makefile compiles with gcc, clang_intel and icc.
Only one flag changes if it is a sandybridge or not.