Fix errors in use of rdtsc
I did some errors in use of rdtsc, fixed before oral presentation.
So in the slides it is correct. In the report, some of the data
in the performance section are wrong, such as the choice of
compiler. But the analysis and conclusions remain unchanged
in the report. I forgot to commit wednesday before the oral.