Modification and tests for a better code in Python
fasta_converter.py converts fasta files in text files with 32 letters
per line (25400 files in 7-8s) for the '\n' in more in string that
indicates the end of 64 bits sequences in binary version.
In moduleDNA.py, I put two versions of read functions, one for text
reader and one for fasta reader.
mymain.py implements convert_to_binary in Python for comparison
with C one, if it's correct in C (~0.00005s for C one, ~0.0015s
for Python one, with 30Ko text file).