Enhancement of convert_to_binary and its Python call
I took the lookup table realised by Guillaume (dev-gui).
I changed almost all. Quickly, I made sure to free the malloc, so
now the function is void type. The parameters in Python are the
string with '\n' character, its size, and the size of future
binary array calculated in Python because easier. And finally, I
have updated the tests in C which are corrects.